% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchdem.R
\name{fetchdem}
\alias{fetchdem}
\title{Fetch and Filter V-Dem Lite Data}
\usage{
fetchdem(
  indicators = NULL,
  start_year = 1970,
  end_year = 2023,
  countries = NULL,
  wide = TRUE
)
}
\arguments{
\item{indicators}{A character vector of indicator names or categories. If NULL, all indicators are included.}

\item{start_year}{The starting year for filtering the data. Default is 1970.}

\item{end_year}{The ending year for filtering the data. Default is 2023.}

\item{countries}{A character vector of country IDs to filter the data. If NULL, all countries are included.}

\item{wide}{Logical, whether to return the data in wide format. If FALSE, returns data in long format. Default is TRUE.}
}
\value{
A data frame containing the filtered V-Dem Lite data in the specified format.
}
\description{
This function fetches and filters data from the V-Dem Lite dataset based on specified indicators, years, and countries.
It can return the data in either wide or long format.
}
\examples{
fetchdem(indicators = c("v2x_polyarchy", "v2x_freexp_altinf"), start_year = 2000, end_year = 2020, countries = c("USA", "SWE"), wide = TRUE)
fetchdem(indicators = "v2x_polyarchy", start_year = 1980, end_year = 2020, countries = NULL, wide = FALSE)
}
