% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchdem.R
\name{search_dem}
\alias{search_dem}
\title{Search and Expand Tags Based on Indices}
\usage{
search_dem(indices = NULL)
}
\arguments{
\item{indices}{A character vector of indices to search and expand. Useful when looking for sub-indices and indicators used to construct a high-level index. Default is \code{NULL}.}
}
\value{
An interactive table with names and descriptions of V-Dem indicators and indices included in the \code{vdemlite} dataset.
}
\description{
This function provides a searchable table of tags and names of of V-Dem indices included in \code{vdemlite}. The table is based on the \href{https://v-dem.net/documents/41/v-dem_structureofaggregation_v14.pdf}{Structure of V-Dem Indices, Components and Indicators} document and includes four columns: the variable tag; the index or indicator descriptor; the level of the index or indicator (high-level index, mid-level index, lower-level index; or indicator); and a column called "Part" that distinguishes between V-Dem Indicators and Indices (part 1) and indices created using V-Dem data (part 2). The default behavior is to return all indices and their descriptions. If a character vector of indices is provided, the function will return any sub-indices and indicators used to construct the selected indices.
}
\examples{
# Return the full table of indices
search_dem()

# Return indices and indicators used to construct polyarchy
search_dem("v2x_polyarchy")

# Return indicators used to construct women's political empowerment and rule of law indices
search_dem(c("v2x_gender", "v2x_rule"))
}
