% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizedem.R
\name{summarize_dem}
\alias{summarize_dem}
\title{Summarize V-Dem Data by Country}
\usage{
summarize_dem(
  indicator = NULL,
  start_year = 1970,
  end_year = 2023,
  countries = NULL,
  visualization = NULL
)
}
\arguments{
\item{indicator}{A character string specifying the V-Dem indicator to summarize.}

\item{start_year}{An integer specifying the start year for the summary (default is 1970).}

\item{end_year}{An integer specifying the end year for the summary (default is 2023).}

\item{countries}{An optional character vector specifying the countries to include (default is NULL, which includes all countries).}

\item{visualization}{An optional character string specifying the type of visualization to include in the summary table (default is NULL, which includes no visualization). Options include "sparkline", "density", or "histogram". Note that it is not recommended to set the visualization parameter if the number of selected countries is large, as it may take a long time to render the table.}
}
\value{
A table summarizing the specified indicator by country.
}
\description{
This function summarizes a specified V-Dem indicator by country for a given time period.
}
\examples{
summarize_dem(indicator = "v2x_polyarchy", start_year = 2000, end_year = 2020)
}
