% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizedem.R
\name{summarizedem}
\alias{summarizedem}
\title{Summarize V-Dem Data by Country}
\usage{
summarizedem(
  indicator = NULL,
  start_year = 1970,
  end_year = 2023,
  countries = NULL
)
}
\arguments{
\item{indicator}{A character string specifying the V-Dem indicator to summarize.}

\item{start_year}{An integer specifying the start year for the summary (default is 1970).}

\item{end_year}{An integer specifying the end year for the summary (default is 2023).}

\item{countries}{An optional character vector specifying the countries to include (default is NULL, which includes all countries).}
}
\value{
A table summarizing the specified indicator by country.
}
\description{
This function summarizes a specified V-Dem indicator by country for a given time period.
}
\examples{
# Summarize the liberal democracy index for all countries and years
summarizedem(indicator = "v2x_libdem")

# Summarize the polyarchy index for all countries for the years 2000-2020
summarizedem(indicator = "v2x_polyarchy", start_year = 2000, end_year = 2020)

# Summarize the civil liberties index for the BRICS with sparklines
countries = c("BRA", "RUS", "IND", "CHN", "ZAF")
summarizedem(indicator = "v2x_civlib",
              countries = countries)

}
